+++ CHANGELOG.md
@@ -6,3 +6,15 @@
 
 - Changed dependency
 - Added delete all instance method
+
+## 0.0.2+1
+
+- avoid overriding
+
+## 0.0.2+2
+
+- fixed deleting all instances
+
+## 0.0.3
+
+- added a flusher to clear instances from memory